% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anomaly.R
\name{anomaly}
\alias{anomaly}
\title{anomaly}
\usage{
anomaly(r, clima)
}
\arguments{
\item{r}{is the Raster layer to compare to the climatology.}

\item{clima}{RasterBrick containing the climatological information.}
}
\description{
This function calculates the anomaly (number of standard
deviations from the mean climatology) of a forecast layer
}
\details{
More information on anomaly is available here:
https://bit.ly/2Qvekz4. You can find fire climatology data for the FWI
indices in Zenodo: https://zenodo.org/communities/wildfire.
}
\examples{
\dontrun{
  r <- brick("cfwis_ffwi_20170101_1200_00.nc")[[1]]
  clima <- brick("fwi.nc")
  x <- anomaly(r, clima)
}

}
