% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/caliver-package.R
\docType{package}
\name{caliver}
\alias{caliver}
\alias{caliver-package}
\title{caliver: CALIbration and VERification of gridded model outputs}
\description{
Utility functions for the post-processing, calibration and validation of gridded model outputs. Initial test cases include the outputs of the following forest fire models: GEFF and RISICO.
}

