% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify_index.R
\name{classify_index}
\alias{classify_index}
\title{classify_index}
\usage{
classify_index(r, thresholds = NULL)
}
\arguments{
\item{r}{is the Raster* object to classify.}

\item{thresholds}{numeric vector containing 5 thresholds.}
}
\value{
The function returns a Raster* object of the same dimensions of
\code{r} but the values are categorical from 1 to 6 (corresponding to
very low, low, moderate, high, very high and extreme danger, respectively).
Some cells may contain NAs, this happens typically over the sea.
}
\description{
This function classifies a fire danger index based on user
defined danger thresholds. If no thresholds are specified, the function
assumes the index is FWI and thresholds are those defined by JRC for Europe:
very Low = 0-5.2, low = 5.2-11.2, moderate = 11.2-21.3, high = 21.3-38.0,
very high = 38-50.0, extreme = > 50.
}
\examples{
\dontrun{
  r <- brick("ECMWF_FWI_20180723_1200_hr_fwi.nc")
  x <- classify_index(r, thresholds = c(5.2, 11.2, 21.3, 38, 50))
}

}
