% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countryMaskCrop.R
\name{countryMaskCrop}
\alias{countryMaskCrop}
\title{Mask a Raster* object using the border of a country}
\usage{
countryMaskCrop(x, countryName, mask = TRUE, crop = TRUE)
}
\arguments{
\item{x}{Raster* to mask}

\item{countryName}{name of the country to use as mask}

\item{mask}{logical (TRUE by default) to decide whether to mask the layer.}

\item{crop}{logical (TRUE by default) to decide whether to crop the layer.}
}
\description{
Mask a Raster* object using the border of a country
}
\details{
The countryName is used to filter spatial polygon with borders from the \code{map} package.
}
\examples{
\dontrun{
  xmasked <- countryMaskCrop(x, countryName = "Spain")
}

}

