% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DangerLevels.R
\name{DangerLevels}
\alias{DangerLevels}
\title{Define Risk Threshold}
\usage{
DangerLevels(fireIndex = "fwi", countryName = "Spain", baseDir = getwd(),
  dataDates, fireSeasonIndex, returnExtremeValues = FALSE)
}
\arguments{
\item{fireIndex}{fire index to calculate the thresholds for (default is fwi = fire weather index)}

\item{countryName}{string describing the country name.}

\item{baseDir}{this is the directory where the reanalysis data are saved.}

\item{dataDates}{dates of the reanalysis data to use.}

\item{fireSeasonIndex}{vector of indices (same length as dataDates) with TRUE if the date falls in the fire season, FALSE otherwise.}

\item{returnExtremeValues}{logical value, by default set to FALSE. If this is set to TRUE, it returns the yearly extreme values (useful to calculate trends).}
}
\value{
A numeric vector listing the thresholds.
}
\description{
This function calculates the danger levels (VeryLow-Low-Moderate-High-VeryHigh-Extreme) for a given country.
}
\examples{
\dontrun{

  baseDir <- "."
  countryName <- "Spain"
  
  # Define period for Reanalysis
  dataDates <- seq.Date(from = as.Date(strptime(paste("1980", 1),
                                                format="\%Y \%j")),
                        to = as.Date(strptime(paste("2016", 366),
                                                format="\%Y \%j")),
                        by = "day")
                        
  # Create an index of fire season dates
  seasons <- getFireSeason(dataDates, 
                           FSS = NULL, FSE = NULL, emisphere = "north")
  fireSeasonIndex <- which(seasons == TRUE)
  
  # Calculate the thresholds
  thresholdFWI <- DangerLevels(fireIndex = "fwi", countryName = "Spain",
                          baseDir, dataDates, fireSeasonIndex)
}

}

