% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{decompress_gz}
\alias{decompress_gz}
\title{decompress_gz (DEPRECATED)}
\usage{
decompress_gz(input_dir = getwd())
}
\arguments{
\item{input_dir}{directory path where all the files to read have been stored}
}
\description{
This function decompresses all .gz files in a given folder
}
\examples{
\dontrun{
  decompress_gz(input_dir = "/var/tmp/moc0/forestfire/")
}

}
