% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decompressGZ.R
\name{decompressGZ}
\alias{decompressGZ}
\title{Decompress all .gz files in a given folder}
\usage{
decompressGZ(dirs, keep = FALSE)
}
\arguments{
\item{dirs}{is the directory where all the files to read are stored}

\item{keep}{logical, if TRUE it keeps the .gz files, removes them otherwise}
}
\description{
This function decompresses all .gz files in a given folder
}
\examples{
\dontrun{
  decompressGZ("/var/tmp/moc0/forestfire/")
}

}

