% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FireDangerLevels.R
\name{FireDangerLevels}
\alias{FireDangerLevels}
\title{Define Risk Threshold}
\usage{
FireDangerLevels(fireIndex, areaOfInterest = NULL, fireSeasonIndex = NULL)
}
\arguments{
\item{fireIndex}{RasterBrick containing the fire index to calculate the thresholds for}

\item{areaOfInterest}{Raster* object or a Spatial* object}

\item{fireSeasonIndex}{index of layers corresponding to the fire season}
}
\value{
A numeric vector listing the thresholds.
}
\description{
This function calculates the danger levels (VeryLow-Low-Moderate-High-VeryHigh-Extreme) for a given country.
}
\examples{
\dontrun{

  # Define period for Reanalysis
  dataDates <- seq.Date(from = as.Date(strptime(paste("1980", 1),
                                                format="\%Y \%j")),
                        to = as.Date(strptime(paste("2016", 366),
                                                format="\%Y \%j")),
                        by = "day")
                        
  # Create an index of fire season dates
  seasons <- getFireSeason(dataDates, emisphere = "north")
  fireSeasonIndex <- which(seasons == TRUE)
  
  # Get area of interest: Europe
  Europe <- getGFED4(varname = 'BasisRegions', region = 'EURO')
  
  # Load and subset FWI
  FWIall <- raster::brick("FWI.nc")
  FWI <- raster::subset(FWIall, fireSeasonIndex)
  
  # Calculate the thresholds
  thresholdFWI <- FireDangerLevels(fireIndex = "fwi", countryName = "Spain",
                          baseDir, dataDates, fireSeasonIndex)
}

}

