% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experimental.R
\name{firegram}
\alias{firegram}
\title{firegram}
\usage{
firegram(thresholds, hres, ens, max_index = 100)
}
\arguments{
\item{thresholds}{is a numeric vector containing the index thresholds
corresponding to: very low, low, moderate, high, very high, extreme danger}

\item{hres}{is the RasterBrick containing the high resolution forecasts}

\item{ens}{is the RasterBrick containing the ensemble forecasts}

\item{max_index}{is the upper limit of the y-axis}
}
\description{
This function plots the distribution of the daily ensemble,
against the hres and previously calculated thresholds
(This function is EXPERIMENTAL, therefore not exported yet!)
}
\examples{
\dontrun{
  p <- firegram(thresholds, hres, ens)
  # Use custom color palette
  p + colorspace::scale_fill_discrete_diverging(palette = "Blue-Red 2")
}

}
