% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experimental.R
\name{forecast_summary_clima}
\alias{forecast_summary_clima}
\title{forecast_summary_clima}
\usage{
forecast_summary_clima(
  input_dir,
  r = NULL,
  p = NULL,
  threshold = 0.95,
  start_date,
  end_date,
  obs = NULL,
  origin = "FWI",
  index = "fwi",
  clima
)
}
\arguments{
\item{input_dir}{folder containing forecast files}

\item{r}{Raster* object containing the forecasts}

\item{p}{SpatialPolygon* identifying the area affected by fires}

\item{threshold}{percentile to be used as danger threshold (default is 0.95).}

\item{start_date}{date when observations start}

\item{end_date}{date when observations end}

\item{obs}{observations, it can either be a file path or a RasterStack/Brick}

\item{origin}{This is the rating system of interest:
fwi (default, currently called cfwis), mark5, nfdrs.}

\item{index}{This is the index to analyse
(default is fwi, belonging to fwi origin).}

\item{clima}{RasterBrick containing the climatological information.}
}
\description{
Plot observations versus forecast using daily climatology
(This function is EXPERIMENTAL, therefore not exported yet!)
}
\examples{
\dontrun{
  forecast_summary(input_dir = "forecast",
                   p = fireBBOX,
                   threshold = 0.95,
                   start_date = "2017-06-01",
                   end_date = "2017-06-30",
                   obs = "CAMS_2017-06-01_2017-06-19_frpfire.nc")
}

}
