% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_fire_danger_levels.R
\name{get_fire_danger_levels}
\alias{get_fire_danger_levels}
\title{get_fire_danger_levels}
\usage{
get_fire_danger_levels(fire_index, ndays = 4)
}
\arguments{
\item{fire_index}{RasterBrick containing the fire index to calculate the
thresholds for. Please note that names(fire_index) should contain dates.}

\item{ndays}{Number of days per year in which a fire is expected to occur.
By default this is 4 days.}
}
\value{
A numeric vector listing the thresholds.
}
\description{
This function calculates the danger levels
(VeryLow-Low-Moderate-High-VeryHigh-Extreme) for a given country.
}
\examples{
\dontrun{

  # Define period for Reanalysis
  dataDates <- seq.Date(from = as.Date(strptime(paste("1980", 1),
                                                format="\%Y \%j")),
                        to = as.Date(strptime(paste("2016", 366),
                                                format="\%Y \%j")),
                        by = "day")

  # Create an index of fire season dates
  seasons <- get_fire_season(dataDates, emisphere = "north")
  fireSeasonIndex <- which(seasons == TRUE)

  r <- raster::brick('FWI_1980-2016.nc')
  p <- raster::getData(name = "GADM", country = "Italy", level = 0)
  maskcrop(r, p, mask = TRUE, crop = TRUE)

  # Mask and crop over country
  FWIcountry <- maskcrop(FWI, country, mask = TRUE, crop = TRUE)

  # Subset based on fire season
  FWIseasonal <- raster::subset(FWIcountry, fireSeasonIndex)

  # Generate thresholds for the country of interest
  countryThr <- get_fire_danger_levels(fire_index = FWIseasonal)

}

}
