% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_fire_season_length.R
\name{get_fire_season_length}
\alias{get_fire_season_length}
\title{get_fire_season_length}
\usage{
get_fire_season_length(x, threshold = 268.15)
}
\arguments{
\item{x}{RasterBrick containing the information.}

\item{threshold}{numeric value used to identify the fire season.}
}
\description{
Get a map of the length of the fire season
(EXPERIMENTAL, not exported yet).
}
\examples{
\dontrun{
  # Load data
  era <- brick("/perm/rd/nen/claudia/era_t2_an_daymax.nc")
  # Calculate map of fire season length
  fire_length <- calc(era, get_fire_season_length)
  # Plot
  plot(rotate(fire_length, main = "Fire season length"))
  maps::map("world", add = TRUE, col = "lightgrey")
  }

}
