% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_gfed4.R
\name{get_gfed4}
\alias{get_gfed4}
\title{get_gfed4}
\usage{
get_gfed4(start_date = NULL, end_date = NULL,
  temporal_resolution = "daily", varname = NULL, region = "GLOB",
  verbose = FALSE)
}
\arguments{
\item{start_date}{first date to download (string)}

\item{end_date}{last date to download (string)}

\item{temporal_resolution}{temporal resolution, it can be "daily" (default)
or "monthly"}

\item{varname}{name of the variable to extract.
This can be one of the following:
\itemize{
  \item{"BasisRegions"}{}
  \item{"BurnedArea"}{}
  \item{"BurnedAreaUncertainty"}{}
  \item{"MeanBurnDateUncertainty"}{}
  \item{"source"}{}
  \item{"TreeCoverDist"}{}
  \item{"LandCoverDist"}{}
  \item{"PeatFraction"}{}
}}

\item{region}{string of characters describing the region.
It can only be one of the 15 values listed below:
\itemize{
  \item{"Global"}{or GLOB}
  \item{"Boreal North America"}{or BONA}
  \item{"Temperate North America"}{or TENA}
  \item{"Central America"}{or CEAM}
  \item{"Northern Hemisphere South America"}{or NHSA}
  \item{"Southern Hemisphere South America"}{or SHSA}
  \item{"Europe"}{or EURO}
  \item{"Middle East"}{or MIDE}
  \item{"Northern Hemisphere Africa"}{or NHAF}
  \item{"Southern Hemisphere Africa"}{or SHAF}
  \item{"Boreal Asia"}{or BOAS}
  \item{"Central Asia"}{or CEAS}
  \item{"Southeast Asia"}{or SEAS}
  \item{"Equatorial Asia"}{or EQAS}
  \item{"Australia and New Zealand"}{or AUST}
}}

\item{verbose}{If TRUE, messages are printed on the screen}
}
\value{
A RasterBrick (if downloading observations) or spatial polygons
(if downloading basis regions)
}
\description{
Get data from the fourth-generation Global Fire Emissions
Database (GFED4)
}
\note{
The conversion from hdf5 to netcdf gets stuck in RStudio
(see \url{https://goo.gl/Rcvp9x}), please use the console.
}
\examples{
\dontrun{

  # Basis regions
  BasisRegions <- get_gfed4(varname = "BasisRegions")

  # Only BONA region
  BONA <- get_gfed4(varname = "BasisRegions", region = "BONA")

  # Monthly burned areas
  BurnedAreas <- get_gfed4(start_date = "2003-01-01",
                           end_date = "2003-01-31",
                           temporal_resolution = "monthly",
                           varname = "BurnedArea")

  # Daily burned areas
  DailyBurnedAreas <- get_gfed4(start_date = "2003-01-01",
                                end_date = "2003-01-02",
                                temporal_resolution = "daily",
                                varname = "BurnedArea")

}

}
