% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_percentile_raster.R
\name{get_percentile_raster}
\alias{get_percentile_raster}
\title{get_percentile_raster}
\usage{
get_percentile_raster(probs, r = NULL, input_file = NULL,
  output_dir = tempdir())
}
\arguments{
\item{probs}{numeric vector of probabilities with values in [0,100] listing
which percentiles should be calculated.}

\item{r}{Raster* object (either RasterStack or RasterBrick).}

\item{input_file}{is the name of the file(path) to read.}

\item{output_dir}{is the directory where the output nc files are saved, by
default this is a temporary directory.}
}
\value{
list containing all the generated percentile maps
}
\description{
This function calculates percentile(s) at each grid point.
}
\examples{
\dontrun{

  x <- get_percentile_raster(probs = c(50, 75, 90, 99),
                             input_file = "./outfile.nc",
                             output_dir = getwd())
}

}
