% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getBurnedAreas.R
\name{getBurnedAreas}
\alias{getBurnedAreas}
\title{Get Burned Areas from the GFED4 database}
\usage{
getBurnedAreas(years, tempRes = "daily", outDir = tempdir(),
  varname = NULL, outFormat = "hdf5", merge = TRUE, keep = FALSE)
}
\arguments{
\item{years}{is a vector containing the years to download}

\item{tempRes}{is the temporal resolution, it can be "daily" (default) or "monthly"}

\item{outDir}{is the directory where the downloaded files are saved, by default this is a temporary directory.}

\item{varname}{name of the variable to extract, by default it keeps all the available variables.}

\item{outFormat}{is the desired format for the output, by default it is "hdf5" but it can also be set equal to "netcdf".}

\item{merge}{logical (TRUE by default). If TRUE it merges all the file over the time dimension.
#@param keep logical (FALSE by default). If FALSE it deletes temporary files.}
}
\description{
This function retrieves Burned Area files from the GFED4 database
}
\note{
The conversion from hdf5 to netcdf gets stuck in RStudio, please use the basic console.
}
\examples{
\dontrun{
  getBurnedAreas(years = 2014, tempRes = "monthly", 
                 outDir = "/var/tmp/moc0/geff/GFED4_BurnedAreas",
                 outFormat = "netcdf", varname = "BurnedArea", 
                 merge = TRUE, keep = FALSE)
}

}

