% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getFireSeason.R
\name{getFireSeason}
\alias{getFireSeason}
\title{Get the season for each date in a vector}
\usage{
getFireSeason(DATES, FSS = NULL, FSE = NULL, zone = "north")
}
\arguments{
\item{DATES}{vector of dates}

\item{FSS}{Fire Season Start (date in the format Y-m-d)}

\item{FSE}{Fire Season End (date in the format Y-m-d)}

\item{zone}{this can either: "north", "south" or "tropics"}
}
\description{
Get the season for each date in a vector
}
\examples{
\dontrun{
  # Modify default seasons
  seasons <- getFireSeason(DATES, 
                           FSS = as.Date("2012-04-01", format = "\%Y-\%m-\%d"), 
                           FSE = as.Date("2012-10-31", format = "\%Y-\%m-\%d"))
}

}
