% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getGEFFreanalysis.R
\name{getGEFFreanalysis}
\alias{getGEFFreanalysis}
\title{getGEFFreanalysis}
\usage{
getGEFFreanalysis(startDate, endDate, varname = "fwi", bb = NULL)
}
\arguments{
\item{startDate}{this is a string with the first date to download, e.g '2015-01-01'}

\item{endDate}{this is a string with the last date to download, e.g '2015-12-31'}

\item{bb}{this is a vector containing the bounding box to get data for. By default the global extent will be downloaded. If a smaller extent is needed, this is expected in the form c(MinLatitude, Maxlatitude, MinLongitude, MaxLongitude).}
}
\value{
A RasterBrick
}
\description{
Get reanalysis data from the Global ECMWF Fire Forecasting (GEFF) system
}
\examples{
\dontrun{

  # Get reanalysis data for 2015 over Italy
  GEFFreanalysis2015 <- getGEFFreanalysis(startDate = '2015-01-01', 
                                          endDate = '2015-12-31',
                                          bb = c(35, 47, 6, 18))
           
}

}

