% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSeason.R
\name{getSeason}
\alias{getSeason}
\title{Get the season for each date in a vector}
\usage{
getSeason(DATES, WS = NULL, SE = NULL, SS = NULL, FE = NULL)
}
\arguments{
\item{DATES}{vector of dates}

\item{WS}{Winter Solstice (date in the format Y-m-d)}

\item{SE}{Spring Equinox (date in the format Y-m-d)}

\item{SS}{Summer Solstice (date in the format Y-m-d)}

\item{FE}{Fall Equinox (date in the format Y-m-d)}
}
\description{
Get the season for each date in a vector
}
\examples{
\dontrun{
  # Modify default seasons
  seasons <- getSeason(DATES, 
                       SS = as.Date("2012-4-1", format = "\%Y-\%m-\%d"), 
                       FE = as.Date("2012-11-1", format = "\%Y-\%m-\%d"))
}

}

