% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getVar.R
\name{getVar}
\alias{getVar}
\title{Extract variable and merge all to one file}
\usage{
getVar(dirs, varname = "fwi")
}
\arguments{
\item{dirs}{is the directory where all the files to read are stored}

\item{varname}{name of the variable to extract}
}
\value{
file containing only the selected variable
}
\description{
This function calculates the CDF at each point of a grid
}
\examples{
\dontrun{
  dirs <- "/var/tmp/moc0/geff"
  x <- getVar(dirs, varname = "fwi")
}

}

