% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caliver-package.R
\docType{data}
\name{GFEDregions}
\alias{GFEDregions}
\title{Data set: GFED BASIS REGIONS}
\format{The format is: Formal class 'RasterLayer' [package "raster"] with 12 slots}
\source{
\url{www.globalfiredata.org/data.html}
}
\usage{
data("GFEDregions")
}
\description{
GFED BASIS REGIONS
}
\keyword{datasets}

