% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chiclet_chart.R
\name{make_chiclet_chart}
\alias{make_chiclet_chart}
\title{make_chiclet_chart}
\usage{
make_chiclet_chart(forecasts, type = "raw", p = NULL, obs = NULL, clima = NULL)
}
\arguments{
\item{forecasts}{either a list of Raster* objects or a string vector
containing forecast file paths. The order of the list/paths should correspond
to forecasts issued consecutively.}

\item{type}{this can be one of the following: "raw" (default), "clima".}

\item{p}{(optional) SpatialPolygon* identifying the area affected by fires}

\item{obs}{(optional) Raster* object containing 1 layer per day of observation}

\item{clima}{(optional) list of Raster* objects containing the climatological information. This is obtained using the function daily_clima() with suitable dates (to cover the full forecasted period) and its extent should contain the polygon `p`, if this is used.}
}
\description{
Generate a summary table of deterministic forecasts.
Example here:
https://news.ucar.edu/9996/long-range-tornado-prediction-it-feasible
}
\examples{
\dontrun{
  forecasts <- c("./ECMWF_FWI_20170101_1200_hr_fwi.nc",
                 "./ECMWF_FWI_20170102_1200_hr_fwi.nc",
                 "./ECMWF_FWI_20170103_1200_hr_fwi.nc",
                 "./ECMWF_FWI_20170104_1200_hr_fwi.nc",
                 "./ECMWF_FWI_20170105_1200_hr_fwi.nc",
                 "./ECMWF_FWI_20170106_1200_hr_fwi.nc",
                 "./ECMWF_FWI_20170107_1200_hr_fwi.nc",
                 "./ECMWF_FWI_20170108_1200_hr_fwi.nc",
                 "./ECMWF_FWI_20170109_1200_hr_fwi.nc",
                 "./ECMWF_FWI_20170110_1200_hr_fwi.nc")
  p <- raster::getData(name = "GADM", country = "IT", level = 0)
  obs <- raster::brick("CAMS_2017-01-01_2017-01-10_frpfire.nc")
  clima <- raster::brick("fwi_1981_2010_Europe.nc")
  df <- make_chiclet_chart(forecasts, p, event_dates, obs, clima)
}

}
