% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast_summary.R
\name{make_forecast_summary}
\alias{make_forecast_summary}
\title{make_forecast_summary}
\usage{
make_forecast_summary(input_dir, p = NULL, event_dates, obs = NULL,
  clima, origin = "FWI", index = "fwi")
}
\arguments{
\item{input_dir}{folder containing forecast files}

\item{p}{SpatialPolygon* identifying the area affected by fires}

\item{event_dates}{date when observations end}

\item{obs}{file path to observations (optional input)}

\item{clima}{file path to dataset containing the climatological information.}

\item{origin}{This is the rating system of interest:
fwi (default, currently called cfwis), mark5, nfdrs.}

\item{index}{This is the index to analyse
(default is fwi, belonging to fwi origin).}
}
\description{
Generate a summary table of deterministic forecasts compared
with daily climatology.
This function is EXPERIMENTAL, therefore not exported yet!
}
\examples{
\dontrun{
  df <- make_forecast_summary(input_dir = "forecast",
                              p = fireBBOX,
                              event_dates = event_dates,
                              obs = "CAMS_frpfire.nc")
}

}
