% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mask_crop_subset.R
\name{mask_crop_subset}
\alias{mask_crop_subset}
\title{mask_crop_subset}
\usage{
mask_crop_subset(r, p, idx = NULL, ...)
}
\arguments{
\item{r}{Raster* object}

\item{p}{SpatialPolygon* object}

\item{idx}{vector of strings indicating the layer indices to subset.}

\item{...}{additional arguments as in writeRaster
(e.g. \code{progress = "text"})}
}
\value{
A Raster* object
}
\description{
mask and/or crop a Raster* based on a Polygon.
}
\details{
Please note that cells along the border with centroids falling
outside the polygon \code{p} will not be returned.
If cells along the border are needed, we suggest to identify cells covering
the polygon and set all remaining pixels to NA, as described in this post:
\url{https://goo.gl/22LwJt}.
}
\examples{
\dontrun{

  r <- raster::brick('FWI_1980-2016.nc')
  p <- raster::getData(name = "GADM", country = "Italy", level = 0)
  mask_crop_subset(r, p)

}

}
