% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maskcrop.R
\name{maskcrop}
\alias{maskcrop}
\title{maskcrop}
\usage{
maskcrop(r, p, mask = TRUE, crop = TRUE)
}
\arguments{
\item{r}{Raster* object}

\item{p}{SpatialPolygon* object}

\item{mask}{logical, TRUE to apply mask (default) using p, FALSE otherwise}

\item{crop}{logical, TRUE to crop (default) using p, FALSE otherwise}
}
\value{
A Raster* object
}
\description{
mask and/or crop a Raster* based on a Polygon.
}
\examples{
\dontrun{

  r <- raster::brick('FWI_1980-2016.nc') 
  p <- raster::getData(name = "GADM", country = "Italy", level = 0)        
  maskcrop(r, p, mask = TRUE, crop = TRUE)
  
}

}

