% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maskcropsub.R
\name{maskcropsub}
\alias{maskcropsub}
\title{maskcrop}
\usage{
maskcropsub(r, p, mask = TRUE, crop = TRUE, i2subset = NULL)
}
\arguments{
\item{r}{Raster* object}

\item{p}{SpatialPolygon* object}

\item{mask}{logical, TRUE to apply mask (default) using p, FALSE otherwise}

\item{crop}{logical, TRUE to crop (default) using p, FALSE otherwise}

\item{i2subset}{vector of strings indicating the layer indices to subset}
}
\value{
A Raster* object
}
\description{
mask and/or crop a Raster* based on a Polygon.
}
\examples{
\dontrun{

  r <- raster::brick('FWI_1980-2016.nc')
  p <- raster::getData(name = "GADM", country = "Italy", level = 0)
  maskcropsub(r, p, mask = TRUE, crop = TRUE)

}

}
