% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean_percs.R
\name{mean_percs}
\alias{mean_percs}
\title{Calculate average percentile}
\usage{
mean_percs(vals, perc.val, mod)
}
\arguments{
\item{vals}{is the a raster object}

\item{perc.val}{is the percentile value used as a threshold}

\item{mod}{defines if the values considered for the mean are above ("gt") or below ("lt") the threshold}
}
\description{
Calculate average percentile
}
