% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mergetime.R
\name{mergetime}
\alias{mergetime}
\title{Merge netcdf files on the time dimension}
\usage{
mergetime(dirs, startingString = "geff_reanalysis_an_fwis_fwi_")
}
\arguments{
\item{dirs}{is the directory where all the files to read are stored}

\item{startingString}{string defining the bginning of the netcdf filenames (needed to exclude other files)}
}
\description{
This function merges all the netcdf files in a given directory over the time dimension. It saves the merged file in the working directory.
}
\examples{
\dontrun{
  # Retrieve mopex daily catalogue
  dirs <- "/var/tmp/moc0/forestfire"
  mergetime(dirs)
}

}

