% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chiclet_chart.R
\name{plot_chiclet_chart}
\alias{plot_chiclet_chart}
\title{plot_chiclet_chart}
\usage{
plot_chiclet_chart(df)
}
\arguments{
\item{df}{data.frame obtained from \code{make_chiclet_chart()}.}
}
\description{
Generate a plot of deterministic forecasts compared
with daily climatology.
}
\examples{
\dontrun{
  df <- make_chiclet_chart(input_dir = "forecast",
                              p = fireBBOX,
                              event_dates = event_dates,
                              obs = "CAMS_frpfire.nc")
  plot_chiclet_chart(df)
}

}
