% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify_index.R
\name{plot_classified_index}
\alias{plot_classified_index}
\title{plot_classified_index}
\usage{
plot_classified_index(r, custom_palette = NULL, ...)
}
\arguments{
\item{r}{is the Raster* object already classified.}

\item{custom_palette}{palette to use (default is \code{viridis::plasma})}

\item{...}{other plotting arguments, see \code{?raster::plot} function.}
}
\description{
Plot classified index as shown in GWIS:
\url{https://bit.ly/2BbBfsm}
}
\examples{
\dontrun{
  r <- brick("cfwis_ffwi_20170101_1200_00.nc")[[1]]
  clima <- brick("fwi.nc")
  anomaly_map <- anomaly(r, clima)
  plot_anomaly(anomaly_map,
               custom_palette = colorRamps::matlab.like(n = length(breaks)))
}

}
