% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_fire_pdf.R
\name{plot_fire_pdf}
\alias{plot_fire_pdf}
\title{plot_fire_pdf}
\usage{
plot_fire_pdf(fire_index, thresholds, upper_limit = NULL, v_lines = NULL)
}
\arguments{
\item{fire_index}{RasterBrick containing the fire index}

\item{thresholds}{thresholds calculated using the function
\code{get_fire_danger_levels()}}

\item{upper_limit}{FWI upper limit to visualise
(the default is the maximum FWI)}

\item{v_lines}{named vector of values to plot as vertical lines
(this can be quantiles for comparison)}
}
\description{
Plot PDF of fire index
}
\examples{
\dontrun{
  plot_fire_pdf(fire_index, thresholds, upper_limit = 100,
                v_lines = c(0.75, 0.95))
}

}
