% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experimental.R
\name{plot_forecast_summary}
\alias{plot_forecast_summary}
\title{plot_forecast_summary}
\usage{
plot_forecast_summary(df)
}
\arguments{
\item{df}{data.frame obtained from \code{make_forecast_summary()}.}
}
\description{
Generate a plot of deterministic forecasts compared
with daily climatology.
This function is EXPERIMENTAL, therefore not exported yet!
}
\examples{
\dontrun{
  df <- make_forecast_summary(input_dir = "forecast",
                              p = fireBBOX,
                              event_dates = event_dates,
                              obs = "CAMS_frpfire.nc")
  plot_forecast_summary(df)
}

}
