% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_obs_vs_forecast.R
\name{plot_obs_vs_forecast}
\alias{plot_obs_vs_forecast}
\title{plot_obs_vs_forecast}
\usage{
plot_obs_vs_forecast(input_dir, s = NULL, p = NULL, threshold,
  start_date, end_date, obs_file_path, forecast_type = "hr",
  origin = "cfwis", index = "ffwi")
}
\arguments{
\item{input_dir}{folder containing forecast files}

\item{s}{Raster* object containing the forecasts}

\item{p}{SpatialPolygon* identifying the area affected by fires.}

\item{threshold}{danger threshold calculated using the function
\code{DangerLevels()}, usually the high danger level.}

\item{start_date}{date when observations start}

\item{end_date}{date when observations end}

\item{obs_file_path}{file path to observations (3D raster)}

\item{forecast_type}{type of forecast: hr (high resolution/deterministic),
ens (control and perturbed forecast/probabilistic)}

\item{origin}{This is the rating system of interest:
fwi (default, currently called cfwis), mark5, nfdrs.}

\item{index}{This is the index to analyse
(default is fwi, belonging to fwi origin).}
}
\description{
Plot observations versus forecast
}
\examples{
\dontrun{
  plot_obs_vs_forecast(input_dir = "/scratch/mo/moc0/fire/GEFF1.2/forecast",
                    p = fireBBOX,
                    threshold = 14,
                    start_date = "2017-06-01", end_date = "2017-06-30",
                    obs_file_path = "CAMS_2017-06-01_2017-06-19_frpfire.nc")
}

}
