% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_percentile_map.R
\name{plot_percentile_map}
\alias{plot_percentile_map}
\title{plot_percentile_map}
\usage{
plot_percentile_map(maps, region = "GLOB", col = NULL, ...)
}
\arguments{
\item{maps}{is the result of get_percentile_map()}

\item{region}{string of characters describing the region.}

\item{col}{custom color palette (default is `dput(rev(RColorBrewer::brewer.pal(n = 10, name = "RdYlGn")))`)}

\item{...}{additional graphical parameters inherited from plot() in the
raster package.}
}
\description{
This function plots the maps of percentiles
}
\examples{
\dontrun{
  # Use default palette
  plot_percentile_map(maps)
  
  # Use custom color palette
  plot_percentile_map(maps, col = rev(viridis::inferno(n = 10)))
}

}
