% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_percentiles.R
\name{plot_percentiles}
\alias{plot_percentiles}
\title{plot_percentiles}
\usage{
plot_percentiles(maps, region = "GLOB", col = NULL, ...)
}
\arguments{
\item{maps}{is the result of get_percentile_raster()}

\item{region}{string of characters describing the region.}

\item{col}{custom color palette (default is `dput(rev(RColorBrewer::brewer.pal(n = 10, name = "RdYlGn")))`)}

\item{...}{additional graphical parameters inherited from plot() in the
raster package.}
}
\description{
This function plots the maps of percentiles
}
\examples{
\dontrun{
  # Use default palette
  plot_percentiles(maps)
  
  # Use custom color palette
  plot_percentiles(maps, col = rev(viridis::inferno(n = 10)))
}

}
