% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotOBSvsForecast.R
\name{plotOBSvsForecast}
\alias{plotOBSvsForecast}
\title{plotOBSvsForecast}
\usage{
plotOBSvsForecast(inDir, areaOfInterest, threshold, startDate, endDate,
  obsFilePath, forecastType = "fc", origin = "fwi", index = "fwi")
}
\arguments{
\item{inDir}{folder containing forecast files}

\item{areaOfInterest}{SpatialPolygon* identifying the area affected by fires.}

\item{threshold}{danger threshold calculated using the function
\code{DangerLevels()}, usually the high danger level.}

\item{startDate}{date when observations start}

\item{endDate}{date when observations end}

\item{obsFilePath}{file path to observations (3D raster)}

\item{origin}{This is the rating system of interest:
fwi (default), mark5, nfdrs.}

\item{index}{This is the index to analyse
(default is fwi, belonging to fwi origin).}
}
\description{
Plot observations versus forecast
}
\examples{
\dontrun{
  plotOBSvsForecast(inDir = "/scratch/mo/moc0/fire/GEFF1.2/forecast", 
                    areaOfInterest = fireBBOX, 
                    threshold = 14, 
                    startDate = "2017-06-01", endDate = "2017-06-30",
                    obsFilePath = "CAMS_2017-06-01_2017-06-19_frpfire.nc")
}

}
