% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/point_inspection.R
\name{point_inspection}
\alias{point_inspection}
\title{point_inspection}
\usage{
point_inspection(
  issue_date,
  points,
  maxleadtime = 10,
  fc_hres = NULL,
  fc_ens = NULL,
  rea_hres = NULL,
  rea_ens = NULL
)
}
\arguments{
\item{issue_date}{x}

\item{points}{x}

\item{maxleadtime}{x}

\item{fc_hres}{x}

\item{fc_ens}{x}

\item{rea_hres}{x}

\item{rea_ens}{x}
}
\value{
A list of named arrays returning the reanalysis and forecast values.
}
\description{
This function inspects reanalysis and forecasts for a given set
of points.
}
\examples{
\dontrun{
  # Before using this function make sure points and all the rasters overlap!
  # Very often points have the longitude in the range -180, +180
  # while rasters have longitudes in the range 0, 360.

  # Define spatial points
  points <- sf::st_as_sf(x = data.frame(long = c(282.88, 171.58, 354.93,
                                                 105.18, 347.52),
                                        lat = c(-12.02, -41.74, 7.73,
                                                -5.1, 22.75)),
                         coords = c("long", "lat"), crs = 4326)
  point_inspection(issue_date = as.Date("2017-01-01"),
                   points,
                   maxleadtime = 10,
                   fc_hres = "/hugetmp/forecasts/hres",
                   fc_ens = "/hugetmp/forecasts/ens",
                   rea_hres = "/hugetmp/reanalysis/hres",
                   rea_ens = "/hugetmp/reanalysis/ens")

}

}
