% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ranking.R
\name{ranking}
\alias{ranking}
\title{ranking}
\usage{
ranking(fc, clima)
}
\arguments{
\item{fc}{is the forecast layer, a Raster* object.}

\item{clima}{RasterBrick containing the climatological information}
}
\description{
This function calculates the percentile ranking of a forecast layer
}
\examples{
\dontrun{
  fc <- brick("cfwis_ffwi_20170101_1200_00.nc")[[1]]
  clima <- brick("fwi.nc")
  x <- ranking(fc, clima)
}

}
