% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readRisicoBinary.R
\name{read_risico_binary}
\alias{read_risico_binary}
\title{read_risico_binary}
\usage{
read_risico_binary(filename)
}
\arguments{
\item{filename}{is the file to read}
}
\description{
Reads a RISICO output file and returns a raster map
}
\examples{
\dontrun{
  r_values <- read_risico_binary("data/RISICO2015_VPPF_201611140300")
}

}
