% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regionalBBOX.R
\name{regionalBBOX}
\alias{regionalBBOX}
\title{Define regional bounding box extent}
\usage{
regionalBBOX(region = "Globe", lonRange = "-180/+180")
}
\arguments{
\item{region}{string of characters describing the region. It can only assume the 12 values listed below:
\itemize{
 \item{"Globe"}{or global} - default extent
 \item{"North America"}{or NAme}
 \item{"Central America"}{or CAme}
 \item{"South America"}{or SAme}
 \item{"Europe"}{or Euro}
 \item{"North Africa"}{or NHAf}
 \item{"South Africa"}{or SHAf}
 \item{"North Asia"}{or NAsi}
 \item{"South Asia"}{or SAsi}
 \item{"Tropical Asia"}{or TAsi}
 \item{"Australia"}{or Aust}
 \item{"East of Moscow"}{or EoMo}
}}

\item{lonRange}{Longitude can be defined in the range "-180/+180" (default) or "0/360".}
}
\description{
This function defines the regional bounding box extent of a map given the name of the region
}
\examples{
\dontrun{
  e <- regionalBBOX(region = "Europe")
}

}

