% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regionalMask.R
\name{regionalMask}
\alias{regionalMask}
\title{Define a mask for the regional extent}
\usage{
regionalMask(region = "GLOB")
}
\arguments{
\item{region}{string of characters describing the region. It can only assume the 15 values listed below:
\itemize{
  \item{"Global"}{or GLOB}
 \item{"Boreal North America"}{or BONA}
 \item{"Temperate North America"}{or TENA}
 \item{"Central America"}{or CEAM}
 \item{"Northern Hemisphere South America"}{or NHSA}
 \item{"Southern Hemisphere South America"}{or SHSA}
 \item{"Europe"}{or EURO}
 \item{"Middle East"}{or MIDE}
 \item{"Northern Hemisphere Africa"}{or NHAF}
 \item{"Southern Hemisphere Africa"}{or SHAF}
 \item{"Boreal Asia"}{or BOAS}
 \item{"Central Asia"}{or CEAS}
 \item{"Southeast Asia"}{or SEAS}
 \item{"Equatorial Asia"}{or EQAS}
 \item{"Australia and New Zealand"}{or AUST}
}}
}
\description{
This function defines the regional extent of a map given the name of the region. It is based on the GFED basis regions defined here: www.globalfiredata.org/data.html
}
\examples{
\dontrun{
  e <- regionalMask(region = "Europe")
}

}

