% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/RiskThresholds.R
\name{RiskThresholds}
\alias{RiskThresholds}
\title{Define Risk Threshold}
\usage{
RiskThresholds(probMaps, regions = "GLOB")
}
\arguments{
\item{probMaps}{is the RasterStack where all the pre-calculated probability maps are stored}

\item{regions}{vector of strings describing the GFED4 region (see regionalMask for options).}
}
\value{
A tibble (data.frame) listing thresholds based on regions (columns) and level of risk (percentile on the rows).
}
\description{
This function calculates the risk threshold given a vector of probabilities corresponding to Low-Medium-High-Extreme risk.
}
\examples{
\dontrun{
  thresholdFWI <- RiskThresholds(probMaps)
}
}

