% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiftMap.R
\name{shiftMap}
\alias{shiftMap}
\title{Shift map}
\usage{
shiftMap(inFile, outFile = NULL)
}
\arguments{
\item{inFile}{input file to be shifted}

\item{outFile}{file path where to save the shifted file}
}
\description{
This function shifts a netcdf file to be in the following bounding box: -180,180-90,90
}
\examples{
\dontrun{
  shift(inFile = "temp.nc", outFile = "output.nc")
}

}

