% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defunct.R
\name{stack_netcdf_files}
\alias{stack_netcdf_files}
\title{Merges all the netcdf files in a given directory over the time dimension}
\usage{
stack_netcdf_files(...)
}
\description{
This function is defunct
}
\keyword{internal}
