% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ThresholdsByRegionR.R
\name{ThresholdsByRegion}
\alias{ThresholdsByRegion}
\title{Validate thresholds}
\usage{
ThresholdsByRegion(obsMerged, reaMerged, region = "EURO", varnames = "FWI",
  MinBurnedArea = 20, probs = 99)
}
\arguments{
\item{obsMerged}{is the nc file name (or path) all the observations are saved, by default this is the file BurnedArea.nc the working directory. Observations can be obtained using the function \code{getBurnedAreas} and then merge them using the function \code{mergetime}.}

\item{reaMerged}{is the nc file name (or path) all the re-analysis are saved, by default this is the file FWI.nc the working directory. This can be generated using the function \code{mergetime}.}

\item{region}{string with the name of the region of interest}

\item{varnames}{string with the name of the fire index}

\item{MinBurnedArea}{Minimum percentage of area burned to consider a cell affected by fire. This is a numeric value in the range [1, 100], by default equal to 20 (20\%).}

\item{probs}{percentile corresponding to the threshold to validate. This is an integer in the range [1, 100], by default equal to 99.}
}
\description{
This function generate a data.frame to compare (and validate) observed burned cells and a fire danger threshold.
}
\examples{
\dontrun{
  library("caliver")
  setwd("/var/tmp/moc0/geff/")
  ThresholdsByRegion(obsMerged = "BurnedArea.nc", 
                     reaMerged = "FWI.nc",
                     region = "EURO", varnames = "FWI", 
                     MinBurnedArea = 20, probs = 99)
}

}

