% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_fire_danger_levels.R
\name{validate_fire_danger_levels}
\alias{validate_fire_danger_levels}
\title{validate_fire_danger_levels}
\usage{
validate_fire_danger_levels(fire_index, observation, fire_threshold,
  obs_threshold)
}
\arguments{
\item{fire_index}{RasterBrick containing the fire index (only one variable)}

\item{observation}{RasterBrick containing the observation (only one variable)}

\item{fire_threshold}{threshold to use to select relevant fire indices}

\item{obs_threshold}{threshold to use to select relevant observations}
}
\value{
A contingency table
}
\description{
This function compares observed and modelled fire data and
return a contingency table summarising the hit rates, false alarms, misses
and correct negatives. The validation can be done using various thresholds
and input data.
}
\examples{
\dontrun{
  setwd("/scratch/mo/moc0/fire/")

  # Load and subset FWI
  BurnedAreas <- raster::brick("GFED4_BurnedAreas/BurnedArea.grd")
  names(BurnedAreas) <- seq.Date(from = as.Date("2003-01-01"),
                                 to = as.Date("2015-12-31"), by = "day")
  FWI <- raster::brick('GEFF/reanalysis/FWI_1980-2016.nc')

  # Generate the contingency table
  ct <- validate_fire_danger_levels(fire_index = FWI,
                                    observation = BurnedAreas,
                                    fire_threshold = 10,
                                    obs_threshold = 50)

}

}
