% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ValidateFireDangerLevels.R
\name{ValidateFireDangerLevels}
\alias{ValidateFireDangerLevels}
\title{Validate Fire Danger Levels}
\usage{
ValidateFireDangerLevels(fireIndex, observation, fireThr, obsThr)
}
\arguments{
\item{fireIndex}{RasterBrick containing the fire index (only one variable)}

\item{observation}{RasterBrick containing the observation (only one variable)}

\item{fireThr}{threshold to use to select relevant fire indices}

\item{obsThr}{threshold to use to select relevant observations}
}
\value{
A contingency table
}
\description{
This function compares observed and modelled fire data and
return a contingency table summarising the hit rates, false alarms, misses
and correct negatives. The validation can be done using various thresholds
and input data.
}
\examples{
\dontrun{
  setwd("/scratch/mo/moc0/fire/")

  # Load and subset FWI
  BurnedAreas <- raster::brick("GFED4_BurnedAreas/BurnedArea.grd")
  names(BurnedAreas) <- seq.Date(from = as.Date("2003-01-01"),
                                 to = as.Date("2015-12-31"), by = "day")
  FWI <- raster::brick('GEFF/reanalysis/FWI_1980-2016.nc')

  # Generate the contingency table
  ct <- ValidateFireDangerLevels(fireIndex = FWI, 
                                 observation = BurnedAreas,
                                 fireThr = 10, obsThr = 50)

}

}
