% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validateThresholds.R
\name{validateThresholds}
\alias{validateThresholds}
\title{Validate thresholds}
\usage{
validateThresholds(obsMerged, riskTable, thresholdBurntArea = 20)
}
\arguments{
\item{obsMerged}{is the nc file name (or path) all the observations are saved, by default this is the file BurnedArea.nc the working directory. Observations can be obtained using the function \code{getBurnedAreas}.}

\item{riskTable}{is the table with all the thresholds}

\item{varnames}{string (or vector of strings) with name of the fire index}
}
\value{
A data.frame
}
\description{
This function allows to validate the thresholds calculated by \code{RiskThresholds}.
}
\examples{
\dontrun{
  library("caliver")
  setwd("/var/tmp/moc0/geff/")
  validateThresholds(obsMerged = "BurnedArea.nc", riskTable = RiskTable)
}

}

