% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anomaly.R
\name{anomaly}
\alias{anomaly}
\title{anomaly}
\usage{
anomaly(fc, clima)
}
\arguments{
\item{fc}{is the forecast layer, a Raster* object.}

\item{clima}{RasterBrick containing the climatological information}
}
\description{
This function calculates the anomaly (standardised deviation from the mean climatology) of a forecast layer
}
\examples{
\dontrun{
  fc <- brick("cfwis_ffwi_20170101_1200_00.nc")[[1]]
  clima <- brick("fwi.nc")
  x <- anomaly(fc, clima)
}

}
