% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anomaly.R
\name{anomaly}
\alias{anomaly}
\title{anomaly}
\usage{
anomaly(r, clima)
}
\arguments{
\item{r}{is the Raster layer to compare to the climatology.}

\item{clima}{RasterBrick containing the climatological information.}
}
\description{
This function calculates the anomaly (standardised deviation
from the mean climatology) of a forecast layer
}
\examples{
\dontrun{
  r <- brick("cfwis_ffwi_20170101_1200_00.nc")[[1]]
  clima <- brick("fwi.nc")
  x <- anomaly(r, clima)
}

}
