% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catNetcdf.R
\name{catNetcdf}
\alias{catNetcdf}
\title{Merge netcdf files on the time dimension}
\usage{
catNetcdf(inDir = NULL, varname = NULL, pattern = NULL,
  recursive = FALSE, outFileName = "outfile.nc", outDir = getwd())
}
\arguments{
\item{inDir}{is the directory where all the files to read are stored}

\item{varname}{name of the variable to extract}

\item{pattern}{regular expression pattern to select a subset of files}

\item{recursive}{logical (TRUE by default). If set to TRUE it looks in folders and subfolders}

\item{outFileName}{output filename}

\item{outDir}{is the directory where outFileName is saved, by default this is the working directory.}
}
\description{
This function merges all the netcdf files in a given directory over the time dimension. It saves the merged file in the working directory.
}
\examples{
\dontrun{
  # Mergetime using single variable nc files
  catNetcdf(inDir = "/var/tmp/moc0/forestfire",
            varname = NULL,
            pattern = "geff_reanalysis_an_fwis_fwi_",
            recursive = TRUE,
            outFileName = "outfile.nc",
            outDir = getwd())
            
}

}

