% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/daily_clima.R
\name{daily_clima}
\alias{daily_clima}
\title{daily_clima}
\usage{
daily_clima(r, dates = NULL)
}
\arguments{
\item{r}{RasterBrick or RasterStack object used to calculate the
climatology. This needs to contain daily layers for multiple years.
\code{names(r)} should also contain dates for comparison
(e.g. X2017.01.01).}

\item{dates}{Dates for which we need to calculate daily climatology.
By default, this is a leap year.}
}
\description{
This function generates daily climatological maps.
}
\examples{
\dontrun{
  daily_clima(r, dates)
}

}
