% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decompressGZ.R
\name{decompressGZ}
\alias{decompressGZ}
\title{decompressGZ}
\usage{
decompressGZ(inDir = getwd())
}
\arguments{
\item{inDir}{directory path where all the files to read have been stored}
}
\description{
This function decompresses all .gz files in a given folder
}
\examples{
\dontrun{
  decompressGZ(inDir = "/var/tmp/moc0/forestfire/")
}

}
