% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decompressGZ.R
\name{decompressGZ}
\alias{decompressGZ}
\title{decompressGZ}
\usage{
decompressGZ(inDir = getwd(), keep = FALSE)
}
\arguments{
\item{inDir}{is the directory path where all the files to read have been previously stored}

\item{keep}{logical, if TRUE it keeps the .gz files, removes them otherwise}
}
\description{
This function decompresses all .gz files in a given folder
}
\examples{
\dontrun{
  decompressGZ(inDir = "/var/tmp/moc0/forestfire/")
}

}

