% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/firegram.R
\name{firegram}
\alias{firegram}
\title{firegram}
\usage{
firegram(thresholds, hres, ens, max_index = 25)
}
\arguments{
\item{thresholds}{is a numeric vector containing the index thresholds
corresponding to: very low, low, moderate, high, very high, extreme danger}

\item{hres}{is the RasterBrick containing the high resolution forecasts}

\item{ens}{is the RasterBrick containing the ensemble forecasts}

\item{max_index}{is the upper limit of the y-axis}
}
\description{
This function plots the distribution of the daily ensemble,
against the hres and previously calculated thresholds
}
\examples{
\dontrun{
  firegram(thresholds, hres, ens)
}

}
