% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_percentile_raster.R
\name{get_percentile_raster}
\alias{get_percentile_raster}
\title{get_percentile_raster}
\usage{
get_percentile_raster(input_file_path, probs, outDir = tempdir())
}
\arguments{
\item{input_file_path}{is the name of the file(path) to read}

\item{probs}{numeric vector of probabilities with values in [0,100] listing
which percentiles should be calculated}

\item{outDir}{is the directory where the output nc files are saved, by
default this is a temporary directory.}
}
\value{
list containing all the generated percentile maps
}
\description{
This function calculates the given percentile at each grid point
}
\examples{
\dontrun{

  x <- get_percentile_raster(input_file_path = "./outfile.nc",
                            probs = c(50, 75, 90, 99),
                            outDir = getwd())
}

}
