% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getBurnedAreas.R
\name{getBurnedAreas}
\alias{getBurnedAreas}
\title{Get Burned Areas from the GFED4 database}
\usage{
getBurnedAreas(years, tempRes = "daily", outDir = getwd())
}
\arguments{
\item{years}{is a vector containing the years to download}

\item{tempRes}{is the temporal resolution, it can be "daily" (default) or "monthly"}

\item{outDir}{is the directory where the downloaded files are saved, by default this is the working directory.}
}
\description{
This function retrieves Burned Area files from the GFED4 database
}
\examples{
\dontrun{
  getBurnedAreas(years = 2014:2015, tempRes = "daily", 
                 outDir = "/var/tmp/moc0/geff/GFED4_BurnedAreas_hdf5/daily")
                 
  getBurnedAreas(years = 2014:2015, tempRes = "monthly", 
                 outDir = "/var/tmp/moc0/geff/GFED4_BurnedAreas_hdf5/monthly")
}

}

