% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getGFED4.R
\name{getGFED4}
\alias{getGFED4}
\title{getGFED4}
\usage{
getGFED4(years = NULL, tempRes = "daily", varname = NULL, merge = TRUE,
  keep = FALSE, region = "GLOB", outDir = getwd(), outFileName = NULL,
  outFormat = "hdf5")
}
\arguments{
\item{years}{is a vector containing the years to download}

\item{tempRes}{is the temporal resolution, it can be "daily" (default) or "monthly"}

\item{varname}{name of the variable to extract, this can be one of the following:
\itemize{
  \item{"BasisRegions"}{}
  \item{"BurnedArea"}{}
  \item{"BurnedAreaUncertainty"}{}
  \item{"MeanBurnDateUncertainty"}{}
  \item{"source"}{}
  \item{"TreeCoverDist"}{}
  \item{"LandCoverDist"}{}
  \item{"PeatFraction"}{}
}}

\item{merge}{logical variable that can either be TRUE or FALSE. If TRUE (default) all the files are merged over the time dimension (this only works if outFormat = "netcdf").}

\item{keep}{logical variable that can either be TRUE or FALSE. If FALSE (default) all the temporary files are deleted (this only works if outFormat = "netcdf").}

\item{region}{string of characters describing the region. It can only assume the 15 values listed below:
\itemize{
  \item{"Global"}{or GLOB}
  \item{"Boreal North America"}{or BONA}
  \item{"Temperate North America"}{or TENA}
  \item{"Central America"}{or CEAM}
  \item{"Northern Hemisphere South America"}{or NHSA}
  \item{"Southern Hemisphere South America"}{or SHSA}
  \item{"Europe"}{or EURO}
  \item{"Middle East"}{or MIDE}
  \item{"Northern Hemisphere Africa"}{or NHAF}
  \item{"Southern Hemisphere Africa"}{or SHAF}
  \item{"Boreal Asia"}{or BOAS}
  \item{"Central Asia"}{or CEAS}
  \item{"Southeast Asia"}{or SEAS}
  \item{"Equatorial Asia"}{or EQAS}
  \item{"Australia and New Zealand"}{or AUST}
}}

\item{outDir}{is the directory where files are saved, by default this is the working directory.}

\item{outFileName}{is the name of the output file (if merged), by default this is the same as the varname.}

\item{outFormat}{is the desired format for the output, by default it is "hdf5" but it can also be set equal to "netcdf".}
}
\description{
Get data from the fourth-generation Global Fire Emissions Database (GFED4)
}
\note{
The conversion from hdf5 to netcdf gets stuck in RStudio, please use the basic console.
}
\examples{
\dontrun{

  # Basis regions
  BasisRegions <- getGFED4(varname = "BasisRegions")

  # Monthly burned areas
  BurnedAreas <- getGFED4(years = 1997:2015, tempRes = "monthly", 
                          varname = "BurnedArea", merge = TRUE, keep = FALSE,
                          outFileName = "BurnedArea_monthly.nc", 
                          outFormat = "netcdf")
                          
  # Daily burned areas
  DailyBurnedAreas <- getGFED4(years = 2003:2017, tempRes = "daily", 
                               varname = "BurnedArea", merge = TRUE, 
                               keep = FALSE, 
                               outFileName = "BurnedArea_daily.nc",
                               outFormat = "netcdf")
           
}

}

