% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getGFED4BasisRegions.R
\name{getGFED4BasisRegions}
\alias{getGFED4BasisRegions}
\title{Get GFED4 Basis Regions}
\usage{
getGFED4BasisRegions(region = NULL, outDir = tempdir())
}
\arguments{
\item{region}{name of the GFED4 basis region.}

\item{outDir}{is the directory where the converted file(s) are saved, by default this is a temporary directory.}
}
\value{
A RasterLayer
}
\description{
This function retrieves the GFED4 basis region from the fourth-generation global fire emissions database website (http://www.globalfiredata.org/data.html).
}
\examples{
\dontrun{
  GFED4BasisRegions <- getGFED4BasisRegions()
}

}

