% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getGriddedCDF.R
\name{getGriddedCDF}
\alias{getGriddedCDF}
\title{Calculate CDF and percentiles}
\usage{
getGriddedCDF(ncfile, probs = c(50, 75, 90, 99), mask = NULL,
  region = "GLOB", outDir = getwd())
}
\arguments{
\item{ncfile}{is the name of the file(path) to read}

\item{probs}{numeric vector of probabilities with values in [0,1] listing which percentiles should be calculated}

\item{mask}{string identifying the name of the mask. By default it is set to NULL, which means no mask is applied. The only mask implemented at the moment is the "fuel_model" provided by JRC (containing non-vegetated areas).}

\item{region}{string identifying the name of the region of interest. By default it is set to provide global coverage ("GLOB") but it can also be used to focus on the reagion listed in \code{regionalMask}.}

\item{outDir}{is the directory where the output nc files are saved, by default this is the working directory.}
}
\value{
list containing all the maps of fwi percentiles
}
\description{
This function calculates the CDF at each point of a grid
}
\examples{
\dontrun{
  ncfile <- "/var/tmp/moc0/forestfire/outfile.nc"
  x <- getGriddedCDF(ncfile)
}

}

