% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getGriddedCDF.R
\name{getGriddedCDF}
\alias{getGriddedCDF}
\title{Calculate CDF}
\usage{
getGriddedCDF(ncfile, varname = "fwi", probs = c(50, 75, 90, 99))
}
\arguments{
\item{ncfile}{is the name of the file to read}

\item{varname}{name of the variable to extract}

\item{probs}{numeric vector of probabilities with values in [0,1] listing which percentiles should be calculated}
}
\value{
list containing all the maps of fwi percentiles
}
\description{
This function calculates the CDF at each point of a grid
}
\examples{
\dontrun{
  # Retrieve mopex daily catalogue
  ncfile <- "/var/tmp/moc0/forestfire/outfile.nc"
  x <- getGriddedCDF(ncfile)
}

}

