% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_GEFF_data.R
\name{import_GEFF_data_from_tar}
\alias{import_GEFF_data_from_tar}
\title{import_GEFF_data_from_tar}
\usage{
import_GEFF_data_from_tar(archive)
}
\arguments{
\item{archive}{file path to the tar file}
}
\description{
This function imports GEFF data (reanalysis or realtime) from a
tar file into a raster stack. Depending on the user's query to the web
portal, the stack can contain different variables.
}
\examples{
\dontrun{
  s <- import_GEFF_data_from_tar(archive = "test.tar")
}

}
