% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{import_geff_data_from_tar}
\alias{import_geff_data_from_tar}
\title{import_geff_data_from_tar (DEPRECATED)}
\usage{
import_geff_data_from_tar(archive, stack_ncfiles = TRUE)
}
\arguments{
\item{archive}{file path to the tar file}

\item{stack_ncfiles}{logical (TRUE by default) variable to decided whether
decompressed nc files should be stacked or not.}
}
\value{
If \code{stack_ncfiles} is TRUE, the function returns a RasterBrick.
If \code{stack_ncfiles} is FALSE, it returns the list of uncompressed files.
}
\description{
This function imports GEFF data (reanalysis or realtime) from a
tar file into a raster stack. Depending on the user's query to the web
portal, the stack can contain different variables.
}
\examples{
\dontrun{
  s <- import_geff_data_from_tar(archive = "test.tar")
}

}
