% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mask_crop_subset.R
\name{mask_crop_subset}
\alias{mask_crop_subset}
\title{mask_crop_subset}
\usage{
mask_crop_subset(r, p, mask = TRUE, crop = TRUE, idx = NULL,
  accurate = FALSE)
}
\arguments{
\item{r}{Raster* object}

\item{p}{SpatialPolygon* object}

\item{mask}{logical, TRUE to apply mask (default) using p, FALSE otherwise}

\item{crop}{logical, TRUE to crop (default) using p, FALSE otherwise}

\item{idx}{vector of strings indicating the layer indices to subset}

\item{accurate}{logical, TRUE to apply more accurate cropping/masking,
FALSE otherwise}
}
\value{
A Raster* object
}
\description{
mask and/or crop a Raster* based on a Polygon.
}
\examples{
\dontrun{

  r <- raster::brick('FWI_1980-2016.nc')
  p <- raster::getData(name = "GADM", country = "Italy", level = 0)
  mask_crop_subset(r, p, mask = TRUE, crop = TRUE)

}

}
