% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mask_with_fuelmodel.R
\name{mask_with_fuelmodel}
\alias{mask_with_fuelmodel}
\title{mask_with_fuelmodel}
\usage{
mask_with_fuelmodel(x)
}
\arguments{
\item{x}{Raster* to mask}
}
\description{
This function masks a Raster* object using JRC's fuelmodel.
}
\details{
In absence of vegetation the risk of ignition reduces considerably,
regardless of the state of the soil. These areas are mapped in the JRC's
fuelmodel. This package contains a cached version of this map. This is stored
in the 'inst/extdata' folder of the package. For evaluating fire danger we
use this map to mask deserts, glaciers, urban areas, etc.
These areas have codes > 20.
}
\examples{
\dontrun{
  xmasked <- mask_with_fuelmodel(x)
}

}
