% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mergetime.R
\name{mergetime}
\alias{mergetime}
\title{Merge netcdf files on the time dimension}
\usage{
mergetime(dirs = NULL, varname = NULL, startingString = "",
  recursive = TRUE, outFile = "outfile.nc", outDir = getwd())
}
\arguments{
\item{dirs}{is the directory where all the files to read are stored}

\item{varname}{name of the variable to extract}

\item{startingString}{string defining the beginning of the netcdf filenames (needed to exclude other files)}

\item{recursive}{logical (TRUE by default). If set to TRUE it looks in folders and subfolders}

\item{outFile}{filename where to store results}

\item{outDir}{is the directory where outFile is saved, by default this is the working directory.}
}
\description{
This function merges all the netcdf files in a given directory over the time dimension. It saves the merged file in the working directory.
}
\examples{
\dontrun{
# Mergetime using single variable nc files
  mergetime(dirs = "/var/tmp/moc0/forestfire",
            varname = NULL,
            startingString = "geff_reanalysis_an_fwis_fwi_",
            recursive = TRUE,
            outFile = "outfile.nc",
            outDir = getwd())
            
}

}

