% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotOBSvsForecast.R
\name{plotOBSvsForecast}
\alias{plotOBSvsForecast}
\title{Plot observations versus forecast}
\usage{
plotOBSvsForecast(inDir, areaOfInterest, threshold, startDate, endDate,
  obsFilePath, origin = "fwi", index = "fwi")
}
\arguments{
\item{inDir}{folder containing forecast files}

\item{areaOfInterest}{SpatialPolygon* identifying the area affected by fires.}

\item{threshold}{danger threshold calculated using the function
\code{DangerLevels()}, usually the high danger level.}

\item{startDate}{FWI upper limit to visualise
(the default is the maximum FWI)}

\item{endDate}{named vector of values to plot as vertical lines
(this can be quantiles for comparison)}

\item{obsFilePath}{file path to observations (3D raster)}

\item{origin}{This is the rating system of interest:
fwi (default), mark5, nfdrs.}

\item{index}{This is the index to analyse
(default is fwi, belonging to fwi origin).}
}
\examples{
\dontrun{
  plotOBSvsForecast(inDir = "/scratch/mo/moc0/fire/GEFF1.2/forecast", 
                    areaOfInterest = fireBBOX, 
                    threshold = 14, 
                    startDate = "2017-06-01", endDate = "2017-06-30",
                    obsFilePath = "/scratch/mo/moc0/fire/events/CAMS_2017-06-01_2017-06-19_frpfire_rotated.nc")
}

}
