% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotOBSvsForecast.R
\name{plotOBSvsForecast}
\alias{plotOBSvsForecast}
\title{Plot observations versus forecast}
\usage{
plotOBSvsForecast(inDir, areaOfInterest, threshold, startDate, endDate,
  obsFilePath)
}
\arguments{
\item{inDir}{folder containing forecast files}

\item{areaOfInterest}{SpatialPolygon* identifying the area affected by fires.}

\item{threshold}{danger threshold calculated using the function \code{DangerLevels()}, usually the high danger level.}

\item{startDate}{FWI upper limit to visualise (the default is the maximum FWI)}

\item{endDate}{named vector of values to plot as vertical lines (this can be quantiles for comparison)}

\item{obsFilePathfile}{path to observations (3D raster)}
}
\description{
Plot observations versus forecast
}
\examples{
\dontrun{
  plotOBSvsForecast(inDir = "/scratch/mo/moc0/fire/GEFF1.2/forecast_fwi", 
                    areaOfInterest = fireBBOX, 
                    threshold = 20, 
                    startDate = "2017-01-01", endDate = "2017-01-31",
                    obsFilePath = "/scratch/mo/moc0/fire/events/CAMS_2017-01-01_2017-01-31_frpfire_rotated.nc")
}

}
