% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPDF.R
\name{plotPDF}
\alias{plotPDF}
\title{Plot PDF of fire index}
\usage{
plotPDF(fireIndex, thresholds, upperLimit = NULL, vLines = NULL)
}
\arguments{
\item{fireIndex}{RasterBrick containing the fire index}

\item{thresholds}{thresholds calculated using the function \code{DangerLevels()}}

\item{upperLimit}{FWI upper limit to visualise (the default is the maximum FWI)}

\item{vLines}{named vector of values to plot as vertical lines (this can be quantiles for comparison)}
}
\description{
Plot PDF of fire index
}
\examples{
\dontrun{
  plotPDF(fireIndex, thresholds, upperLimit = 100)
}

}
