% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPercentiles.R
\name{plotPercentiles}
\alias{plotPercentiles}
\title{Plot maps of percentiles}
\usage{
plotPercentiles(listOfMaps, backgroundMap = FALSE)
}
\arguments{
\item{listOfMaps}{is the result of getGriddedCDF()}

\item{backgroundMap}{logical, if TRUE it uses a background map}
}
\description{
This function plots the maps of percentiles contained in x
}
\examples{
\dontrun{
  plotPercentiles(listOfMaps)
}

}

