% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/plotPercentiles.R
\name{plotPercentiles}
\alias{plotPercentiles}
\title{Plot maps of percentiles}
\usage{
plotPercentiles(maps, rotateMap = FALSE, region = "GLOB", ...)
}
\arguments{
\item{maps}{is the result of getGriddedCDF()}

\item{rotateMap}{logical, if TRUE it uses a background map}

\item{region}{string of characters describing the region.}

\item{...}{additional graphical parameters inherited from plot() in the raster package.}
}
\description{
This function plots the maps of percentiles
}
\examples{
\dontrun{
  plotPercentiles(maps)
}
}

