% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPercentiles.R
\name{plotPercentiles}
\alias{plotPercentiles}
\title{Plot maps of percentiles}
\usage{
plotPercentiles(listOfMaps)
}
\arguments{
\item{listOfMaps}{is the result of getGriddedCDF()}
}
\description{
This function plots the maps of percentiles contained in x
}
\examples{
\dontrun{
  plotPercentiles(listOfMaps)
}

}

