% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ranking.R
\name{ranking}
\alias{ranking}
\title{ranking}
\usage{
ranking(r, clima)
}
\arguments{
\item{r}{is the RasterLayer to compare to the climatology.}

\item{clima}{RasterBrick containing the climatological information.}
}
\description{
The ranking is applied to a forecast map \code{r} and provides
percentiles of occurrence of the values based on a given climatology
(see \code{clima}).
}
\details{
More information on ranking is available here:
https://bit.ly/2Qvekz4. To estimate fire climatology one can use hindcast or
reanalysis data. Examples of the latter are available from Zenodo:
https://zenodo.org/communities/wildfire.
}
\examples{
\dontrun{
  r <- brick("cfwis_ffwi_20170101_1200_00.nc")[[1]]
  clima <- brick("fwi.nc")
  x <- ranking(r, clima)
  
  # This plots nicely using rasterVis::levelplot(), in example on GWIS 
  # (\\url{https://gwis.jrc.ec.europa.eu}
  rasterVis::levelplot(x, col.regions = colorRamps::matlab.like(n = 6))
  rasterVis::levelplot(x, col.regions = c("green", "yellow", "salmon",
                                          "orange", "red", "black"))
}

}
