% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_risico_binary.R
\name{read_risico_binary}
\alias{read_risico_binary}
\title{read_risico_binary}
\usage{
read_risico_binary(filename)
}
\arguments{
\item{filename}{is the file to read}
}
\value{
The function returns a RasterLayer or Brick, depending on whether
\code{filenane} contains one or more layers.
}
\description{
Reads a RISICO output file and returns a raster map
}
\examples{
\dontrun{
  read_risico_binary(system.file("extdata", "RISICO_binary.bin", package = "caliver"))
}

}
