% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readRisicoBinary.R
\name{readRisicoBinary}
\alias{readRisicoBinary}
\title{Read a RISICO output file and return a raster map}
\usage{
readRisicoBinary(filename)
}
\arguments{
\item{filename}{is the file to read}
}
\description{
Reads a RISICO output file and returns a raster map
}
\examples{
\dontrun{
  r_values <- readRisicoBinary("data/RISICO2015_VPPF_201611140300")
}

}
