% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regionalBBOX.R
\name{regionalBBOX}
\alias{regionalBBOX}
\title{Define regional bounding box extent}
\usage{
regionalBBOX(region = "Globe", lonRange = "0/360")
}
\arguments{
\item{region}{string of characters describing the region. It can only assume the 15 values listed below:
\itemize{
  \item{"Global"}{or GLOB}
 \item{"Boreal North America"}{or BONA}
 \item{"Temperate North America"}{or TENA}
 \item{"Central America"}{or CEAM}
 \item{"Northern Hemisphere South America"}{or NHSA}
 \item{"Southern Hemisphere South America"}{or SHSA}
 \item{"Europe"}{or EURO}
 \item{"Middle East"}{or MIDE}
 \item{"Northern Hemisphere Africa"}{or NHAF}
 \item{"Southern Hemisphere Africa"}{or SHAF}
 \item{"Boreal Asia"}{or BOAS}
 \item{"Central Asia"}{or CEAS}
 \item{"Southeast Asia"}{or SEAS}
 \item{"Equatorial Asia"}{or EQAS}
 \item{"Australia and New Zealand"}{or AUST}
}}

\item{lonRange}{Longitude can be defined in the range "-180/+180" (default) or "0/360".}
}
\description{
This function defines the regional bounding box extent of a map given the name of the region
}
\examples{
\dontrun{
  e <- regionalBBOX(region = "Europe")
}

}

