% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiftMap.R
\name{shiftMap}
\alias{shiftMap}
\title{Shift map}
\usage{
shiftMap(inFile)
}
\arguments{
\item{inFile}{input file(s) to be shifted}
}
\value{
A shifted file for ech input file name
}
\description{
This function shifts a netcdf file to be in the following bounding box: -180,180-90,90
}
\examples{
\dontrun{
  shift(inFile = "temp.nc", outFile = "output.nc")
}

}

