% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stack_with_rat.R
\name{stack_with_rat}
\alias{stack_with_rat}
\title{stack_with_rat}
\usage{
stack_with_rat(r, ids, classes)
}
\arguments{
\item{r}{is the categorical Raster* object.}

\item{ids}{numeric vector containing the categorical levels.}

\item{classes}{character vector containing the categorical labels.}
}
\value{
The function returns a RasterStack of the same dimensions of
\code{r} but with associated rat.
}
\description{
This function converts the Raster* object that contains
categorical values to a stack and associates a Raster Attribute Table (rat).
This is useful if plotting using the \code{rasterVis::levelplot()} function.
}
\examples{
\dontrun{
  x <- stack_with_rat(r, ids = 1:3, classes = c("Low", "Moderate", "High")
}

}
